refactor(home): home section 컴포넌트 파일명 변경 및 통일
