+++ domains/home/ui/index.ts
@@ -1,6 +1,8 @@
 export * from './sheets';
 export * from './states';
 export * from './todo';
-export * from './user';
+export * from './section';
 
+export * from './habit-streak-card';
+export * from './user-profile-card';
 export * from './calendar-section';
