+++ domains/home/ui/section/habit-streak-section.tsx
@@ -1,13 +1,12 @@
 import { useMemo } from 'react';
 
 import { useTodosWithSuspense } from '@/domains/home/hooks';
+import { HabitStreakCard } from '@/domains/home/ui';
 import { getContinuousDays } from '@/domains/home/utils';
 import dayjs from '@/shared/lib/dayjs';
 import { QueryErrorBoundary } from '@/shared/ui/async-boundary';
 
-import { HabitStreakCard } from './habit-streak-card';
-
-const HabitStreakContent = () => {
+const HabitStreakSectionContent = () => {
   const yearMonth = dayjs().format('YYYY-MM');
   const { data } = useTodosWithSuspense(yearMonth);
 
@@ -27,13 +26,13 @@ const HabitStreakContent = () => {
   );
 };
 
-export const HabitStreak = () => {
+export const HabitStreakSection = () => {
   return (
     <QueryErrorBoundary
       loadingFallback={<HabitStreakCard.Skeleton />}
       errorFallback={<HabitStreakCard.Skeleton />}
     >
-      <HabitStreakContent />
+      <HabitStreakSectionContent />
     </QueryErrorBoundary>
   );
 };
