+++ domains/home/ui/section/index.ts
@@ -1,0 +1,2 @@
+export * from './habit-streak-section';
+export * from './user-greeting-section';
