--- domains/home/ui/user/index.ts
@@ -1,2 +1,0 @@
-export * from './user-greeting';
-export * from './habit-streak';
